setup/init/requirements.txt
argon2-cffi>=21.3.0
psycopg2-binary>=2.9.9
typer>=0.15.2
